+++ gax-grpc/src/main/java/com/google/api/gax/grpc/CallOptionsUtil.java
@@ -29,24 +29,44 @@
  */
 package com.google.api.gax.grpc;
 
-import static org.junit.Assert.assertEquals;
-
 import com.google.common.collect.ImmutableMap;
+import io.grpc.CallOptions;
+import io.grpc.Metadata;
 import io.grpc.Metadata.Key;
+import java.util.Collections;
 import java.util.Map;
-import org.junit.Test;
 
-public class GrpcCallContextTest {
-  @Test
-  public void testWithRequestParamsDynamicHeaderOption() {
-    String encodedRequestParams = "param1=value&param2.param3=value23";
-    GrpcCallContext context =
-        GrpcCallContext.of().withRequestParamsDynamicHeaderOption(encodedRequestParams);
+/** A utility class that provides helper functions to work with custom call options. */
+class CallOptionsUtil {
+  // this is a call option name, not a header name, it is not transferred over the wire
+  private static final CallOptions.Key<Map<Key<String>, String>> DYNAMIC_HEADERS_CALL_OPTION_KEY =
+      CallOptions.Key.of("gax_dynamic_headers", Collections.<Key<String>, String>emptyMap());
+  // this is the header name, it is transferred over the wire
+  static Metadata.Key<String> REQUEST_PARAMS_HEADER_KEY =
+      Metadata.Key.of("x-goog-request-params", Metadata.ASCII_STRING_MARSHALLER);
 
-    Map<Key<String>, String> headers =
-        CallOptionsUtil.getDynamicHeadersOption(context.getCallOptions());
+  private CallOptionsUtil() {}
 
-    assertEquals(
-        ImmutableMap.of(CallOptionsUtil.REQUEST_PARAMS_HEADER_KEY, encodedRequestParams), headers);
+  static CallOptions putRequestParamsDynamicHeaderOption(
+      CallOptions callOptions, String requestParams) {
+    if (callOptions == null || requestParams.isEmpty()) {
+      return callOptions;
+    }
+
+    Map<Key<String>, String> dynamicHeadersOption =
+        callOptions.getOption(DYNAMIC_HEADERS_CALL_OPTION_KEY);
+
+    // This will fail, if REQUEST_PARAMS_HEADER_KEY is already there
+    dynamicHeadersOption =
+        ImmutableMap.<Key<String>, String>builder()
+            .putAll(dynamicHeadersOption)
+            .put(REQUEST_PARAMS_HEADER_KEY, requestParams)
+            .build();
+
+    return callOptions.withOption(DYNAMIC_HEADERS_CALL_OPTION_KEY, dynamicHeadersOption);
+  }
+
+  static Map<Key<String>, String> getDynamicHeadersOption(CallOptions callOptions) {
+    return callOptions.getOption(DYNAMIC_HEADERS_CALL_OPTION_KEY);
   }
 }
